+++ aerogear-android-authz-test/src/main/java/org/jboss/aerogear/android/authorization/test/helper/Data.java
@@ -14,9 +14,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.aerogear.android.impl.helper;
+package org.jboss.aerogear.android.authorization.test.helper;
 
-import org.jboss.aerogear.android.RecordId;
+import org.jboss.aerogear.android.core.RecordId;
 
 public class Data implements Comparable<Data> {
