+++ aerogear-android-authz-test/src/main/java/org/jboss/aerogear/android/authorization/test/util/PatchedActivityInstrumentationTestCase.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.aerogear.android.impl.util;
+package org.jboss.aerogear.android.authorization.test.util;
 
 import android.app.Activity;
 import android.test.ActivityInstrumentationTestCase2;
