+++ aerogear-android-authz/src/main/java/org/jboss/aerogear/android/authorization/oauth2/OAuth2AuthroizationConfigurationProvider.java
@@ -14,16 +14,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.aerogear.android.impl.util;
+package org.jboss.aerogear.android.authorization.oauth2;
 
-import android.app.Activity;
-import android.os.Bundle;
+import org.jboss.aerogear.android.core.ConfigurationProvider;
 
-public class StubActivity extends Activity {
+public class OAuth2AuthroizationConfigurationProvider implements ConfigurationProvider<OAuth2AuthorizationConfiguration> {
 
     @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
+    public OAuth2AuthorizationConfiguration newConfiguration() {
+        return new OAuth2AuthorizationConfiguration();
     }
 
 }
