+++ aerogear-android-authz/src/main/java/org/jboss/aerogear/android/authorization/oauth2/OAuth2AuthzSession.java
@@ -14,11 +14,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.aerogear.android.impl.authz.oauth2;
+package org.jboss.aerogear.android.authorization.oauth2;
 
 import android.os.Parcel;
 import android.os.Parcelable;
-import org.jboss.aerogear.android.RecordId;
+import org.jboss.aerogear.android.core.RecordId;
 
 import java.util.Date;
